extract getSignableMessage to utils
